lcd and keypad kernel module added

+++ LCD_platform_driver/lcd_platform_driver_data.h
@@ -15,18 +15,15 @@ struct lcd_gpio_dev_data{
 };
 
 
-
-
 /*holds lcd device specific data*/
 struct lcd_dev_data {
 
     char text[30];
     int x;
     int y;
-    int cmd;
+    long cmd;
 
 };
-
 
 
 /*holds driver specific data*/
